Added 'Convert to Multilingual' option (2.6 only)
This features allows developer to quickly convert a Text Box field into a Multilingual Field in a click of a button.

It copies values and settings and the field_id is kept.
If the original field is required, all languages are set has required.
If the original field is optional, no languages are required.

To use this feature, you must use Symphony 2.6.x and up.
If you run an earlier version, the button is hidden.