Generated handle can be bigger than 255 chars
Under mysql 5.6+, the strict mode is activated by default, which changes
the behavior from auto-truncate to an error when a varchar overflows.
The current code allowed for bigger than 255 chars handle creation, when
the created handle is 254 or 255 characters. The code appended two more
chars if that handle is already being used, resulting in an overflow.

This commit makes sure we always truncate the created handle to create
space for the "-count" part.

This is a port of
https://github.com/symphonists/textboxfield/commit/c46fadf7c2ddf8841bc72f6cd2106e87fd72b839