Fix problem where All languages gets forced
The problem was that when saving a section with errors, the All
languages option would become selected.

We need the 'required' property to be properly set to be able to call
the parent properly. But we need to "revert" its value after we are done
with it. We must also check for 'all' when creating the UI, not the
'required' property.

Fixes #82